+++ app/integrations/facetwp/class-facetwp.php
@@ -27,6 +27,19 @@ if ( ! defined( 'ABSPATH' ) ) {
 class FacetWP extends Integration implements Registerable, Service {
 
 	/**
+	 * Additional functional files.
+	 *
+	 * @var array $helper_files
+	 * @since 2.2.0
+	 */
+	protected $helper_files = [
+		'template-hooks',
+		'template-functions',
+		'default-facets',
+		'customize',
+	];
+
+	/**
 	 * Additional inline CSS configuration items.
 	 *
 	 * @var array $inline_css_configs
@@ -42,12 +55,9 @@ class FacetWP extends Integration implements Registerable, Service {
 	 * @since 1.0.0
 	 */
 	public function register() {
-		add_filter( 'bigbox_customize_inline_css_configs', [ $this, 'inline_css_configs' ] );
+		$this->load_helper_files();
 
-		include_once $this->get_dir() . '/template-hooks.php';
-		include_once $this->get_dir() . '/template-functions.php';
-		include_once $this->get_dir() . '/default-facets.php';
-		include_once $this->get_dir() . '/customize.php';
+		add_filter( 'bigbox_customize_inline_css_configs', [ $this, 'inline_css_configs' ] );
 	}
 
 }