+++ app/integrations/woocommerce/template-hooks.php
@@ -35,6 +35,9 @@ add_filter( 'bigbox_navbar_search', 'bigbox_woocommerce_navbar_search' );
  * File: archive-product.php.
  */
 
+// Remove sidebar on single product pages.
+add_action( 'wp', 'bigbox_woocommerce_template_sidebars' );
+
 // Replace outer content wrapper.
 remove_action( 'woocommerce_before_main_content', 'woocommerce_output_content_wrapper' );
 remove_action( 'woocommerce_after_main_content', 'woocommerce_output_content_wrapper_end' );
@@ -99,9 +102,6 @@ add_action( 'woocommerce_after_shop_loop_item_title', 'bigbox_woocommerce_templa
 /**
  * File: content-single-product.php.
  */
-
-// Remove sidebar on single product pages.
-add_action( 'wp', 'bigbox_woocommerce_template_tertiary' );
 
 // Filter product tabs.
 add_filter( 'woocommerce_product_tabs', 'bigbox_woocommerce_product_tabs', 20 );