+++ bin/generate-google-font-list.sh
@@ -3,6 +3,10 @@
 # Include useful functions
 source "$(dirname "$0")/wp-bin/wp-bin.sh"
 
-DATA=$(download "https://www.googleapis.com/webfonts/v1/webfonts?key=AIzaSyBmZBJAajbmRv6fSyfN3Dh5iz5XkWqYjDs")
+read -s -p "Enter your Google Fonts API key: " key
 
-echo ${DATA} | jq '.items | map({family, category, variants, subsets})' > "./resources/data/google-fonts.json"
+if [ "$key" ]
+then
+	DATA=$(download "https://www.googleapis.com/webfonts/v1/webfonts?key=$key")
+	echo ${DATA} | jq '.items | map({family, category, variants, subsets})' > "./resources/data/google-fonts.json"
+fi