+++ resources/assets/js/app/navbar.js
@@ -1,3 +1,5 @@
+/* global _ */
+
 /**
  * External dependencies.
  */