+++ resources/assets/js/license-manager/models/license.js
@@ -1,4 +1,4 @@
-/* global Backbone */
+/* global BigBoxLicenseManager, Backbone */
 
 /**
  * External dependencies.